+++ aksel.nav.no/website/components/layout/header/search/search.module.css
@@ -1,0 +1,34 @@
+.searchGrid {
+  display: grid;
+  grid-template-columns: 12rem auto;
+  grid-template-areas:
+    "input input"
+    "results results";
+}
+
+.searchGrid .searchFilter {
+  display: none;
+}
+
+.searchGrid .searchInput {
+  grid-area: input;
+}
+
+.searchGrid .searchResults {
+  grid-area: results;
+}
+
+@media (min-width: 768px) {
+  .searchGrid {
+    display: grid;
+    grid-template-columns: 12rem auto;
+    grid-template-areas:
+      "empty input"
+      "filter results";
+  }
+
+  .searchGrid .searchFilter {
+    display: block;
+    grid-area: filter;
+  }
+}
