+++ aksel.nav.no/website/components/layout/header/search/useDebounce.ts
@@ -1,0 +1,13 @@
+import { useEffect, useState } from "react";
+
+export const useDebounce = (value: any) => {
+  const [debouncedValue, setDebouncedValue] = useState(value);
+
+  useEffect(() => {
+    const timeOut = setTimeout(() => setDebouncedValue(value), 1000);
+
+    return () => clearTimeout(timeOut);
+  }, [value]);
+
+  return debouncedValue;
+};
